cloud_server_python/requirements.txt
# Python package requirements for Cloud Server (CS)

protobuf
python-dotenv

# To install these requirements:
# pip install -r requirements.txt

# For reproducibility, it's good practice to pin versions.
# After installing and ensuring everything works, you can generate
# a requirements file with specific versions using:
# pip freeze > requirements.txt
# Example of pinned versions (your versions might differ):
# protobuf==3.20.1
# python-dotenv==0.21.0